==> 04 numeric_score_regression.ipynb <==
import pandas as pd
import statsmodels.api as sm
import statsmodels.formula.api as smf

dat = pd.read_csv("data/train_test_set2.csv")
dat

dat.describe()

dat.columns

predictors = [
    "CODE_GENDER",
    "FLAG_OWN_CAR",
    "FLAG_OWN_REALTY",
    "CNT_CHILDREN",
    "AMT_INCOME_TOTAL",
    "NAME_INCOME_TYPE",
    "NAME_EDUCATION_TYPE",
    "NAME_FAMILY_STATUS",
    "NAME_HOUSING_TYPE",
    "DAYS_BIRTH",
    "DAYS_EMPLOYED",
    "FLAG_MOBIL",
    "FLAG_WORK_PHONE",
    "FLAG_PHONE",
    "FLAG_EMAIL",
    "OCCUPATION_TYPE",
    "CNT_FAM_MEMBERS",
    "AGE"
]

targets = [
    'NORM_SUM',
    'LOG_SUM',
    'NORM_MEAN',
    'LOG_MEAN',
    'MIN_MAX_LOG_SUM',
    'MIN_MAX_LOG_MEAN',
    'MIN_MAX_NORM_SUM',
    'MIN_MAX_NORM_MEAN',
    'MIN_MAX_SUM',
    'MIN_MAX_MEAN'
]

for y in targets:
    r2s = list()
    print(f"Target = {y}")
    print("----------------------------")
    for x in predictors:
        res = smf.ols(f"{y} ~ {x}", dat).fit()
        r2s.append(res.rsquared)
        print(f"R2 using {x} = {res.rsquared}")
    print("----------------------------")
    print(f"Best R2 for {y}: {max(r2s)}")
    print("----------------------------\n")

The maximum percentage of variance accounted for by any application field for any version of the score is 0.43%, which indicates that the variance in the scores is more likely to be due to some interaction between the variables.  Decision trees are a great way to automatically take interactions between application fields into account.

from sklearn.tree import DecisionTreeRegressor
from sklearn.model_selection import train_test_split
import numpy as np

train, test = train_test_split(dat)
train

test_X = pd.get_dummies(test.loc[:, predictors])
test_ys = test.loc[:, targets]
train_X = pd.get_dummies(train.loc[:, predictors])
train_ys = train.loc[:, targets]

rmses = list()
for target in targets:
    train_y = train_ys.loc[:, target]
    test_y = test_ys.loc[:, target]
    dt = DecisionTreeRegressor()
    dt.fit(train_X, train_y)
    preds = dt.predict(test_X)
    actuals = test_y
    model_rmse = np.sqrt(((preds - actuals)**2).mean())
    nullhypo_rmse = np.sqrt(((train_y.mean() - actuals)**2).mean())
    print(
        f"NullHypo RMSE for target {y}: {nullhypo_rmse}",
        f"Model RMSE for target {y}: {model_rmse}",
        f"-----------------------------",
        sep="\n"
    )

For every version of the numeric score, the decision tree regressor performs worse than the null hypothesis.  We will thus need a different approach to reliability than any of these numeric scores.
