==> 07 leaf_node_EDA.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

leaf_nodes = [
    pd.read_pickle(f"leaf_node_pkls/node_{i}.pkl")
    for i in range(839)
]

min(node.shape[0] for node in leaf_nodes)

max(node.shape[0] for node in leaf_nodes)

node_pmfs = [
    pd.Series(0, index=np.arange(-6, 2))
    for node in leaf_nodes
]
for idx, node in enumerate(leaf_nodes):
    curr_pmf = node["STATUS"].value_counts() / node.shape[0]
    node_pmfs[idx][curr_pmf.index] = curr_pmf
print(node_pmfs)

mins = {
    i: min(
        pmf[i] 
        for pmf in node_pmfs
    )
    for i in range(-6, 2)
}
maxs = {
    i: max(
        pmf[i] 
        for pmf in node_pmfs
    )
    for i in range(-6, 2)
}

mins

maxs

for idx, pmf in enumerate(node_pmfs):
    plt.bar(pmf.index, pmf.values)
    plt.ylim((0,1))
    plt.title(f"PMF of STATUS for Cluster {idx}")
    plt.savefig(f"leaf_node_plots_w_titles/Node {idx}")
    plt.close()

zeroless_leaf_nodes = [
    pd.read_pickle(f"leaf_nodes_no_zeros_pkls/zeroless_node_{i}.pkl")
    for i in range(703)
]

min(node.shape[0] for node in zeroless_leaf_nodes), max(node.shape[0] for node in zeroless_leaf_nodes)

zeroless_node_pmfs = [
    pd.Series(0, index=[-6,-5,-4,-3,-2,-1,1])
    for node in zeroless_leaf_nodes
]
for idx, node in enumerate(zeroless_leaf_nodes):
    curr_pmf = node["STATUS"].value_counts() / node.shape[0]
    zeroless_node_pmfs[idx][curr_pmf.index] = curr_pmf
print(zeroless_node_pmfs)

mins = {
    i: min(
        pmf[i] 
        for pmf in zeroless_node_pmfs
    )
    for i in (-6,-5,-4,-3,-2,-1,1)
}
maxs = {
    i: max(
        pmf[i] 
        for pmf in zeroless_node_pmfs
    )
    for i in (-6,-5,-4,-3,-2,-1,1)
}

mins, maxs

for idx, pmf in enumerate(zeroless_node_pmfs):
    plt.bar(["-6", "-5", "-4", "-3", "-2", "-1", "1"], pmf.values)
    plt.ylim((0,1))
    plt.savefig(f"zeroless_leaf_node_plots2/ZerolessNode {idx}")
    plt.close()

zeroless_train = pd.read_pickle("data/train_no_zeros.pkl")
pmf = zeroless_train["STATUS"].value_counts() / zeroless_train.shape[0]
pmf.sort_index(inplace=True)
plt.bar(["-6", "-5", "-4", "-3", "-2", "-1", "1"], pmf, color="red")
plt.ylim((0,1))
plt.title("PMF of Zeroless Training Set STATUS")
plt.xlabel("STATUS")
plt.savefig(f"zeroless_leaf_node_plots2/Full training set")
plt.close()
