==> 08 sklearn_DecisionTreeRegressor.ipynb <==
import numpy as np
import pandas as pd
from sklearn.tree import DecisionTreeRegressor

train = pd.read_pickle("data/train.pkl")
test = pd.read_pickle("data/test_by_ID.pkl")

def calculate_pmf(df, idx):
    pmf = pd.Series(0, index=idx)
    counts = df["STATUS"].value_counts()
    pmf[counts.index] = counts / df.shape[0]
    return pmf

# get PMF by ID for training sets
statuses = np.arange(-6,2)
train_by_ID = train.groupby("ID")


pmfs = pd.DataFrame(
    [
        calculate_pmf(group, idx=statuses)
        for _, group in train_by_ID
    ],
    index=set(train["ID"])
)
pmfs

pmfs.index.name = "ID"
pmfs

len(set(train["ID"])), pmfs.shape

predictors = [
    "CODE_GENDER",
    "FLAG_OWN_CAR",
    "FLAG_OWN_REALTY",
    "CNT_CHILDREN",
    "AMT_INCOME_TOTAL",
    "NAME_INCOME_TYPE",
    "NAME_EDUCATION_TYPE",
    "NAME_FAMILY_STATUS",
    "NAME_HOUSING_TYPE",
    "DAYS_BIRTH",
    "DAYS_EMPLOYED",
    "FLAG_MOBIL",
    "FLAG_WORK_PHONE",
    "FLAG_PHONE",
    "FLAG_EMAIL",
    "OCCUPATION_TYPE",
    "CNT_FAM_MEMBERS"
]

train_by_ID = train[["ID", "STATUS"]+predictors].drop_duplicates("ID")
train_by_ID.set_index("ID", inplace=True)
train_by_ID

dtr_byid = DecisionTreeRegressor()
dtr_byid.fit(X=pd.get_dummies(train_by_ID.loc[:, predictors]), y=pmfs)

actual_pmfs = pd.read_pickle("model_evaluation/actual_pmfs.pkl").values
pred_pmfs = dtr_byid.predict(pd.get_dummies(test[predictors]))
pred_pmfs_normed = pred_pmfs / np.sum(pred_pmfs)
model_msqe = np.sqrt(np.mean((pred_pmfs - actual_pmfs)**2, axis=0))

nullhypo_preds = pd.Series(0, index=statuses)
pcts = train_by_ID["STATUS"].value_counts() / train_by_ID.shape[0]
nullhypo_preds[pcts.index] = pcts
nullhypo_preds.sort_index(inplace=True)
nullhypo_msqe = np.sqrt(np.mean((nullhypo_preds.values - actual_pmfs)**2, axis=0))

model_msqe, nullhypo_msqe

Regressor consistently worse than the null hypothesis when training by ID.  Let's try letting each month count as its own row instead.

for status in statuses:
    train[status] = 0.0
train.set_index("ID", inplace=True)

train.loc[:, statuses] = pmfs
train

train.to_pickle("data/train_w_pmfs.pkl")

dtr = DecisionTreeRegressor()
dtr.fit(X=pd.get_dummies(train.loc[:, predictors]), y=train.loc[:, statuses])

pred_pmfs = dtr.predict(pd.get_dummies(test[predictors]))
pred_pmfs_normed = pred_pmfs / np.sum(pred_pmfs)
actual_pmfs = pd.read_pickle("model_evaluation/actual_pmfs.pkl").values
model_msqe = np.sqrt(np.mean((pred_pmfs - actual_pmfs)**2, axis=0))

nullhypo_preds = pd.Series(0, index=statuses)
pcts = train["STATUS"].value_counts() / train.shape[0]
nullhypo_preds[pcts.index] = pcts
nullhypo_msqe = np.sqrt(np.mean((nullhypo_preds.values - actual_pmfs)**2, axis=0))

model_msqe, nullhypo_msqe

The regression model still performs worse than the null hypothesis.

Try a list of XGBoostRegressors, one for each status.

from xgboost import XGBRegressor

xg_regressors = {
    target_status: XGBRegressor(n_estimators=10)
    for target_status in range(-6,2)
}

for target_status, regressor in xg_regressors.items():
    print("training...")
    regressor.fit(X=pd.get_dummies(train.loc[:, predictors]), y=train.loc[:, target_status])

predictions = np.array([
    regressor.predict(pd.get_dummies(test[predictors]))
    for regressor in xg_regressors.values()
]).T

np.sqrt(np.mean((predictions - actual_pmfs)**2, axis=0))

This also performs worse than the null hypothesis.

from sklearn.ensemble import GradientBoostingRegressor

gbr = GradientBoostingRegressor()
gbr.fit()
